requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
neo4j==4.1.0
neotime==1.0.0
numpy==1.19.3
pytz==2020.1
six==1.15.0
Werkzeug==0.16.0
uWSGI==2.0.17.1
python-dotenv==0.15.0
